Refactor code to abstract the rendered button
