+++ src/components/RenderButton/index.js
@@ -1,0 +1,3 @@
+import RenderButton from './RenderButton';
+
+export default RenderButton;
